feat: no collect for controllers
