+++ src/libs/satellite/src/usage/mod.rs
@@ -2,3 +2,4 @@ mod impls;
 mod store;
 pub mod types;
 pub mod user_usage;
+pub mod assert;
